requirements.txt
# requirements.txt
# Requires Python >= 3.9

manim==0.19.0
langchain==0.3.23
langgraph==0.3.25
langchain-google-genai==2.1.2
python-dotenv==1.1.0

# Testing
pytest==8.3.2
pytest-mock==3.14.0
